feat: implement spec 203 - fix duplicate DebtType enums
Consolidates duplicate DebtType enum definitions into a single source of truth.

Changes:
- Removed unit variant DebtType enum from src/core/mod.rs
- Now use only struct variant enum from src/priority/mod.rs (re-exported from core)
- Added missing legacy variants (Todo, Fixme, CodeSmell, etc.) to priority::DebtType
- Added PartialEq, Eq, Hash trait implementations with custom Hash for f64 fields
- Added Display implementations for DebtType and ErrorSwallowingPattern
- Updated all 50+ files to use struct variant syntax with { .. } patterns
- Fixed all detectors to create struct variants with appropriate field values
- Updated categorize_debt_type to handle all struct variants
- Fixed suppression logic to compare variant types using discriminants
- Updated all tests (400+ tests) to use struct variant syntax

Modified files: 50+ across analyzers, debt detectors, tests, and supporting modules

Spec: specs/203-fix-duplicate-debttype-enums.md (deleted after implementation)

+++ benches/tier_classification_bench.rs
@@ -3,12 +3,13 @@
 //! Validates that tier classification has acceptable performance
 //! and ensures no regression after refactoring to smaller functions.
 
-use criterion::{black_box, criterion_group, criterion_main, BenchmarkId, Criterion};
+use criterion::{criterion_group, criterion_main, BenchmarkId, Criterion};
 use debtmap::priority::tiers::pure::classify_tier;
 use debtmap::priority::{
     ActionableRecommendation, DebtType, FunctionRole, ImpactMetrics, Location, TierConfig,
     UnifiedDebtItem, UnifiedScore,
 };
+use std::hint::black_box;
 use std::path::PathBuf;
 
 fn create_test_item(
@@ -266,8 +267,8 @@ fn bench_classify_batch(c: &mut Criterion) {
             .map(|i| {
                 let debt_type = match i % 4 {
                     0 => DebtType::GodObject {
-                        methods: 50 + (i % 50) as usize,
-                        fields: 20 + (i % 30) as usize,
+                        methods: 50 + (i % 50),
+                        fields: 20 + (i % 30),
                         responsibilities: 5,
                         god_object_score: 85.0,
                     },
